aegis-pm/requirements.txt
# AutoGen
pyautogen==0.2.35

# FastAPI
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP / Integrations
httpx==0.27.0
openai==1.35.0

# Scheduling
apscheduler==3.10.4

# Retry logic
tenacity==8.3.0

# Config
python-dotenv==1.0.1
pydantic-settings==2.3.3

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
aiosqlite==0.20.0
